Fix compatibility for macOS 14.0 and higher
The new code doesn't use GUI scripting, as it reveals the specific area which was previously reached by clicking the button.
The main benefit is that if it does not find that section, the script will simply fail, instead of changing the wrong user setting.

Of course, it assumes that the pane id is known, as this could be different on other OS versions. 

The script could be improved further by relying solely on these ```reveal``` calls for navigation.